Change nvim current line from 0 to actual line.
Still keep the relative numbers for the other lines as it helps traverse
the opened file. I concur that you can see the current line in the
status bar, but it can eventually be modified to not display the current
line, and my eyes are up on the line anyways.

Also, add automatic "}" when inserting "{" into the file. It
simplifies the writing of code in C, which I am learning at the moment.
(It will help any other language that uses { })