Add readme with commands for dot migration
Use https://www.atlassian.com/git/tutorials/dotfiles as reference; my
commands are different and for this reason the new file is a README and
not a shell script.

Finally, and this is not reflected anywhere, interactive rebase is used
to change the email I used for the previous commits. I had the wrong
config. Annoyingly so, it reset my dates and I had to approximate the
hours of the commits.

Both "dot rebase -i --root" and "GIT_COMMITTER_DATE= dot commit -amend"
were used. Couldn't imagine doing this in a project with multiple
programmers though!