Add a printf map in nvim
It may not be the best implementation, I may need to change it later
down the line, but it works for me now.

It's nice to not have to write "); everytime I'm done writing what to
output. Now yes, in the case of writing for example "%d\n", x); it could
bite me in the ass but again, I'll have to see.