Remove history file
The only time I'd be looking at my cmd history would be if it was a once
in a blue moon esoteric cmd that I found on a backwater www for the most
specific of problems. At which point, it'd be my fault for not
safeguarding it like the forbidden spell it is.