Replace init.vim with init.lua
This was arduous. Not because of the translation, but because of Alpine.
At first, my init.lua was invisble to nvim. I tried reinstalling both
nvim and lua, and then I gave up and used the `-u` flag.

The 'require' keyword, which links other files/libs to the file,
usually, from what I understood, automatically finds any .lua written
inside lua/ directories. Well lo and behold as this installation of nvim
(or is the issue lua) can't find the .config/nvim/lua directory.

So I'll just do everything in one config file. Deal with it future me!